Reset Alembic version
