--- alembic/versions/4b41206a16ea_add_task_and_backgroundjob.py
@@ -1,61 +1,0 @@
-"""Add Task and BackgroundJob
-
-Revision ID: 4b41206a16ea
-Revises: ce4d216a6d0b
-Create Date: 2020-04-18 14:20:14.591991
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = '4b41206a16ea'
-down_revision = 'ce4d216a6d0b'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table(
-        'task',
-        sa.Column('id', sa.Integer(), nullable=False),
-        sa.Column('name', sa.String(length=128), nullable=False),
-        sa.Column('default_params', sa.JSON(), nullable=False),
-        sa.Column('created_at', sa.DateTime(timezone=True),
-                  server_default=sa.text('(CURRENT_TIMESTAMP)'), nullable=True),
-        sa.Column('updated_at', sa.DateTime(
-            timezone=True), nullable=True),
-        sa.PrimaryKeyConstraint('id')
-    )
-    op.create_table(
-        'background_job',
-        sa.Column('id', sa.Integer(), nullable=False),
-        sa.Column('type', sa.String(length=64), nullable=False),
-        sa.Column('params', sa.JSON(), nullable=False),
-        sa.Column('output', sa.JSON(), nullable=False),
-        sa.Column('status', sa.String(length=64), nullable=False),
-        sa.Column('created_at', sa.DateTime(timezone=True),
-                  server_default=sa.text('(CURRENT_TIMESTAMP)'), nullable=True),
-        sa.Column('updated_at', sa.DateTime(
-            timezone=True), nullable=True),
-        sa.Column('task_id', sa.Integer(), nullable=True),
-        sa.ForeignKeyConstraint(['task_id'], ['task.id'], ),
-        sa.PrimaryKeyConstraint('id')
-    )
-    with op.batch_alter_table('background_job', schema=None) as batch_op:
-        batch_op.create_index(batch_op.f('ix_background_job_type'), [
-                              'type'], unique=False)
-
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    with op.batch_alter_table('background_job', schema=None) as batch_op:
-        batch_op.drop_index(batch_op.f('ix_background_job_type'))
-
-    op.drop_table('background_job')
-    op.drop_table('task')
-    # ### end Alembic commands ###
