--- alembic/versions/8a771bc3c9eb_remove_model_data_add_model_config_and_.py
@@ -1,41 +1,0 @@
-"""Remove Model.data, add Model.config and a foreign key to ClassificationTrainingData
-
-Revision ID: 8a771bc3c9eb
-Revises: d61ca505f5b1
-Create Date: 2020-04-29 15:56:34.197265
-
-"""
-from alembic import op
-import sqlalchemy as sa
-from sqlalchemy.dialects import sqlite
-
-# revision identifiers, used by Alembic.
-revision = '8a771bc3c9eb'
-down_revision = 'd61ca505f5b1'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    with op.batch_alter_table('model', schema=None) as batch_op:
-        batch_op.add_column(
-            sa.Column('classification_training_data_id', sa.Integer(), nullable=True))
-        batch_op.add_column(sa.Column('config', sa.JSON(), nullable=True))
-        batch_op.create_foreign_key(batch_op.f('fk_model_classification_training_data_id_classification_training_data'),
-                                    'classification_training_data', ['classification_training_data_id'], ['id'])
-        batch_op.drop_column('data')
-
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    with op.batch_alter_table('model', schema=None) as batch_op:
-        batch_op.add_column(sa.Column('data', sqlite.JSON(), nullable=False))
-        batch_op.drop_constraint(batch_op.f(
-            'fk_model_classification_training_data_id_classification_training_data'), type_='foreignkey')
-        batch_op.drop_column('config')
-        batch_op.drop_column('classification_training_data_id')
-
-    # ### end Alembic commands ###
