--- alembic/versions/9b304b3f69dd_rename_annotation_table_to_.py
@@ -1,67 +1,0 @@
-"""rename annotation table to classification_annotation
-
-Revision ID: 9b304b3f69dd
-Revises: 4b41206a16ea
-Create Date: 2020-04-20 14:37:12.976765
-
-"""
-import sqlalchemy as sa
-from alembic import op
-
-# revision identifiers, used by Alembic.
-revision = '9b304b3f69dd'
-down_revision = '4b41206a16ea'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('classification_annotation',
-                    sa.Column('id', sa.Integer(), nullable=False),
-                    sa.Column('value', sa.Integer(), nullable=False),
-                    sa.Column('created_at', sa.DateTime(timezone=True),
-                              server_default=sa.text('(CURRENT_TIMESTAMP)'),
-                              nullable=True),
-                    sa.Column('updated_at', sa.DateTime(timezone=True),
-                              nullable=True),
-                    sa.Column('entity_id', sa.Integer(), nullable=True),
-                    sa.Column('label_id', sa.Integer(), nullable=True),
-                    sa.Column('user_id', sa.Integer(), nullable=True),
-                    sa.Column('context_id', sa.Integer(), nullable=True),
-                    sa.ForeignKeyConstraint(['context_id'], ['context.id'], ),
-                    sa.ForeignKeyConstraint(['entity_id'], ['entity.id'], ),
-                    sa.ForeignKeyConstraint(['label_id'], ['label.id'], ),
-                    sa.ForeignKeyConstraint(['user_id'], ['user.id'], ),
-                    sa.PrimaryKeyConstraint('id')
-                    )
-    # TODO we may need to think about copying over data from the annotation
-    #  table to the classification_annotation. We are OK for now since there
-    #  are no data in the table anyway.
-    #  In case we need some guidance:
-    #  https://stackoverflow.com/questions/18726527/fetch-table-values-using-alembic-and-update-to-another-table 
-    op.drop_table('annotation')
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('annotation',
-                    sa.Column('id', sa.INTEGER(), nullable=False),
-                    sa.Column('value', sa.INTEGER(), nullable=False),
-                    sa.Column('created_at', sa.DATETIME(), nullable=False),
-                    sa.Column('last_updated_at', sa.DATETIME(),
-                              nullable=False),
-                    sa.Column('entity_id', sa.INTEGER(), nullable=True),
-                    sa.Column('label_id', sa.INTEGER(), nullable=True),
-                    sa.Column('user_id', sa.INTEGER(), nullable=True),
-                    sa.Column('context_id', sa.INTEGER(), nullable=True),
-                    sa.ForeignKeyConstraint(['context_id'], ['context.id'], ),
-                    sa.ForeignKeyConstraint(['entity_id'], ['entity.id'], ),
-                    sa.ForeignKeyConstraint(['label_id'], ['label.id'], ),
-                    sa.ForeignKeyConstraint(['user_id'], ['user.id'], ),
-                    sa.PrimaryKeyConstraint('id')
-                    )
-    # TODO the reverse of the data copy in the upgrade().
-    op.drop_table('classification_annotation')
-    # ### end Alembic commands ###
