--- alembic/versions/a59bd0856f5c_tweak_model_columns.py
@@ -1,58 +1,0 @@
-"""tweak model columns
-
-Revision ID: a59bd0856f5c
-Revises: db7129d57819
-Create Date: 2020-04-24 08:39:40.623419
-
-"""
-from alembic import op
-import sqlalchemy as sa
-from sqlalchemy.dialects import sqlite
-
-# revision identifiers, used by Alembic.
-revision = 'a59bd0856f5c'
-down_revision = 'db7129d57819'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    with op.batch_alter_table('classification_training_data', schema=None) as batch_op:
-        batch_op.drop_column('output_filename')
-
-    with op.batch_alter_table('file_inference', schema=None) as batch_op:
-        batch_op.drop_column('output_filename')
-
-    with op.batch_alter_table('model', schema=None) as batch_op:
-        batch_op.add_column(sa.Column('data', sa.JSON(), nullable=False))
-        batch_op.add_column(sa.Column('uuid', sa.String(length=64), nullable=False))
-        batch_op.add_column(sa.Column('version', sa.Integer(), nullable=False))
-        batch_op.create_unique_constraint('_uuid_version_uc', ['uuid', 'version'])
-        batch_op.create_index(batch_op.f('ix_model_uuid'), ['uuid'], unique=False)
-        batch_op.create_index(batch_op.f('ix_model_version'), ['version'], unique=False)
-        batch_op.drop_column('params')
-        batch_op.drop_column('output')
-
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    with op.batch_alter_table('model', schema=None) as batch_op:
-        batch_op.add_column(sa.Column('output', sqlite.JSON(), nullable=False))
-        batch_op.add_column(sa.Column('params', sqlite.JSON(), nullable=False))
-        batch_op.drop_index(batch_op.f('ix_model_version'))
-        batch_op.drop_index(batch_op.f('ix_model_uuid'))
-        batch_op.drop_constraint('_uuid_version_uc', type_='unique')
-        batch_op.drop_column('version')
-        batch_op.drop_column('uuid')
-        batch_op.drop_column('data')
-
-    with op.batch_alter_table('file_inference', schema=None) as batch_op:
-        batch_op.add_column(sa.Column('output_filename', sa.TEXT(), nullable=False))
-
-    with op.batch_alter_table('classification_training_data', schema=None) as batch_op:
-        batch_op.add_column(sa.Column('output_filename', sa.TEXT(), nullable=False))
-
-    # ### end Alembic commands ###
