--- alembic/versions/b3796ff2df5a_add_annotationrequest.py
@@ -1,56 +1,0 @@
-"""Add AnnotationRequest
-
-Revision ID: b3796ff2df5a
-Revises: 9b304b3f69dd
-Create Date: 2020-04-20 22:25:39.575625
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = 'b3796ff2df5a'
-down_revision = '9b304b3f69dd'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table(
-        'annotation_request',
-        sa.Column('id', sa.Integer(), nullable=False),
-        sa.Column('created_at', sa.DateTime(timezone=True), server_default=sa.text(
-            '(CURRENT_TIMESTAMP)'), nullable=True),
-        sa.Column('user_id', sa.Integer(), nullable=False),
-        sa.Column('context_id', sa.Integer(), nullable=False),
-        sa.Column('annotation_type', sa.Integer(), nullable=False),
-        sa.Column('status', sa.Integer(), nullable=False),
-        sa.Column('task_id', sa.Integer(), nullable=True),
-        sa.Column('order', sa.Float(), nullable=True),
-        sa.Column('name', sa.String(), nullable=True),
-        sa.Column('additional_info', sa.JSON(), nullable=True),
-        sa.Column('source', sa.JSON(), nullable=True),
-        sa.ForeignKeyConstraint(['context_id'], ['context.id'], ),
-        sa.ForeignKeyConstraint(['task_id'], ['task.id'], ),
-        sa.ForeignKeyConstraint(['user_id'], ['user.id'], ),
-        sa.PrimaryKeyConstraint('id')
-    )
-    with op.batch_alter_table('annotation_request', schema=None) as batch_op:
-        batch_op.create_index(batch_op.f('ix_annotation_request_order'), [
-                              'order'], unique=False)
-        batch_op.create_index(batch_op.f('ix_annotation_request_status'), [
-                              'status'], unique=False)
-
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    with op.batch_alter_table('annotation_request', schema=None) as batch_op:
-        batch_op.drop_index(batch_op.f('ix_annotation_request_status'))
-        batch_op.drop_index(batch_op.f('ix_annotation_request_order'))
-
-    op.drop_table('annotation_request')
-    # ### end Alembic commands ###
