--- alembic/versions/bf2c53c6acc4_add_classification_annotation_id_to_.py
@@ -1,34 +1,0 @@
-"""add classification annotation id to annotation request
-
-Revision ID: bf2c53c6acc4
-Revises: a59bd0856f5c
-Create Date: 2020-04-24 17:35:30.405208
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = 'bf2c53c6acc4'
-down_revision = 'a59bd0856f5c'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    with op.batch_alter_table('annotation_request', schema=None) as batch_op:
-        batch_op.add_column(sa.Column('classification_annotation_id', sa.Integer(), nullable=True))
-        batch_op.create_foreign_key(batch_op.f('fk_annotation_request_classification_annotation_id_classification_annotation'), 'classification_annotation', ['classification_annotation_id'], ['id'])
-
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    with op.batch_alter_table('annotation_request', schema=None) as batch_op:
-        batch_op.drop_constraint(batch_op.f('fk_annotation_request_classification_annotation_id_classification_annotation'), type_='foreignkey')
-        batch_op.drop_column('classification_annotation_id')
-
-    # ### end Alembic commands ###
